done Home Page animation!!!
