+++ packages/static-ui/src/containers/Home/index.js
@@ -1,26 +1,16 @@
 import React, {Component} from "react";
-import root from "react-shadow";
-import Webflow from "./webflow";
-import __html from "./landing_page";
-import landingPageAnimation from "./landing_page.animation";
+import __html from "./static_page";
+import startAnimation from "./animation";
 
 class LandingPage extends Component {
   componentDidMount = () => {
-    landingPageAnimation()
+    startAnimation()
   };
 
   render() {
     return (
-        // <root.div id="shadow-dom" mode="open">
-        //   <Webflow />
-          <html
-              style={{
-                all: 'initial'
-              }}
-              data-wf-page="60cacdc34d2a0e6df119b3b2"
-              dangerouslySetInnerHTML={{__html}}
-          />
-        // </root.div>
+        <div style={{all: 'initial'}}
+             dangerouslySetInnerHTML={{__html}}/>
     );
   }
 }
