--- packages/static-ui/src/containers/Home/reset-style.css
@@ -1,3 +1,0 @@
-:host {
-    all: initial
-}
