--- packages/static-ui/src/containers/Home/webflow.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import __html from "./landing_page";
-import landingPageAnimation from "./landing_page.animation";
-
-class Webflow extends React.Component {
-    // componentDidMount = () => {
-    //     landingPageAnimation()
-    // };
-    render() {
-        return (
-            // <div id="shadow-dom" dangerouslySetInnerHTML={{__html}}/>
-            <div dangerouslySetInnerHTML={{__html}}/>
-        );
-    }
-}
-
-export default Webflow;
